Foundational changes for future update
The idea for some of these changes is to allow for caching of profile pictures on the local system. Right now when changing between pages; react will load new versions of the users current profile picture when in reality it should only really load changes if they've been modified. This will save on bandwidth.